requirements.txt
pytesseract
pdf2image
pillow
python-docx
transformers
torch
